third_party/WebKit/LayoutTests/editing/pasteboard/paste-text-at-tabspan-002-expected.txt

EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification

You should see a\tax\t\tz below:
| "
"
| <span>
|   id="test"
|   "a"
|   <span>
|     class="Apple-tab-span"
|     style="white-space:pre"
|     "	"
|   "ax<#selection-caret>"
|   <span>
|     class="Apple-tab-span"
|     style="white-space:pre"
|     "		"
|   "z"
| "
"
